+++ src/MicroElements.Metadata/Metadata/IPropertyValueFactory.cs
@@ -1,0 +1,20 @@
+﻿// Copyright (c) MicroElements. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace MicroElements.Metadata
+{
+    /// <summary>
+    /// Represents <see cref="IPropertyValue"/> factory.
+    /// </summary>
+    public interface IPropertyValueFactory
+    {
+        /// <summary>
+        /// Creates new instance if <see cref="IPropertyValue"/>.
+        /// </summary>
+        /// <param name="property">Property.</param>
+        /// <param name="value">Value for property.</param>
+        /// <param name="valueSource">Value source.</param>
+        /// <returns>Created property value.</returns>
+        IPropertyValue Create(IProperty property, object? value, ValueSource? valueSource = null);
+    }
+}