+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-# 7.0.0-rc.20
+# 7.0.0-rc.22
 - Performance optimization for metadata search
+- IPropertyValueFactory, CachedPropertyValueFactory
+- XmlParser: IPropertyValueFactory added to IXmlParserSettings
+- CachedStringParser and InernedStringParser
 
 # 7.0.0-rc.19
 - Validation for XmlParser